requirements.txt
viam-sdk
ultralytics
ultralyticsplus